# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LeadForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: Get all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/search:
    post:
      operationId: searchLeads
      tags: [leads]
      summary: Search/find leads globally
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadSearchInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadSearchResult"
  /leads/import:
    post:
      operationId: importLeads
      tags: [leads]
      summary: Import leads (CSV or bulk)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadImportInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadImportResult"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        status:
          type: string
        score:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, status, createdAt, updatedAt]
    LeadInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        title:
          type: string
        website:
          type: string
        country:
          type: string
        city:
          type: string
        industry:
          type: string
        status:
          type: string
        score:
          type: integer
        notes:
          type: string
        linkedinUrl:
          type: string
        source:
          type: string
      required: [name]
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        title:
          type: string
        website:
          type: string
        country:
          type: string
        city:
          type: string
        industry:
          type: string
        status:
          type: string
        score:
          type: integer
        notes:
          type: string
        linkedinUrl:
          type: string
        source:
          type: string
    LeadSearchInput:
      type: object
      properties:
        query:
          type: string
        industry:
          type: string
        country:
          type: string
        title:
          type: string
        companySize:
          type: string
        limit:
          type: integer
      required: [query]
    LeadSearchResult:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/LeadCandidate"
        total:
          type: integer
        source:
          type: string
      required: [leads, total, source]
    LeadCandidate:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        confidence:
          type: integer
      required: [name, confidence]
    LeadImportInput:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/LeadInput"
      required: [leads]
    LeadImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [imported, skipped, errors]
...