# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LeadForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadUpdate {
  name?: string;
  email?: string;
  phone?: string;
  company?: string;
  title?: string;
  website?: string;
  country?: string;
  city?: string;
  industry?: string;
  status?: string;
  score?: number;
  notes?: string;
  linkedinUrl?: string;
  source?: string;
}